ciao-controller: Implement instance migration
This commit adds support for instance migration in controller.  It does
this my mapping the stop command received from clients to an SSNTP
DELETE command and the restart command to an SSNTP start.  There are
ofcourse some complications.  A special flag, migration, is set in the
SSNTP DELETE command payload to indicate that the instance being deleted
is actually being migrated.  This causes launcher to send an InstanceStopped
event, rather than an InstanceDeleted event, when it has deleted the
instance from the local node.  Controller has been modified to remove
the association between a node and an instance and to set the instance
state to exited when it receives an InstanceStopped event, but to otherwise
leave the instance untouched.  Launcher also sends InstanceStopped events when
an instance dies of its own accord, e.g., its a container whose workload has
finished.

At a later stage when the user issues a restart command controller now
recreates the original START command from the information it stores about
the instance.  This command is sent to scheduler and the instance is
restarted on a new node.

Fixes #1121
Fixes #1043

+++ ciao-controller/internal/datastore/datastore.go
@@ -907,6 +907,11 @@ func (ds *Datastore) StopFailure(instanceID string, reason payloads.StopFailureR
 // for this tenant. If the instance was a normal tenant instance, the
 // IP address will be released and the instance will be deleted from the
 // datastore.
+//
+// Only instances whose status is pending are removed when a StartFailure event
+// is received.  StartFailure errors may also be generated when restarting an
+// exited instance and we want to make sure that a failure to restart such
+// an instance does not result in it being deleted.
 func (ds *Datastore) StartFailure(instanceID string, reason payloads.StartFailureReason) error {
 	var tenantID string
 	var cnci bool
@@ -952,7 +957,7 @@ func (ds *Datastore) StartFailure(instanceID string, reason payloads.StartFailur
 		return errors.Wrapf(err, "error getting instance (%v)", instanceID)
 	}
 
-	if reason.IsFatal() {
+	if reason.IsFatal() && (i.State == payloads.ComputeStatusPending) {
 		ds.deleteInstance(instanceID)
 	}
 
@@ -1035,7 +1040,9 @@ func (ds *Datastore) deleteInstance(instanceID string) (string, error) {
 
 	ds.tenantsLock.Lock()
 	tenant := ds.tenants[i.TenantID]
-	delete(tenant.instances, instanceID)
+	if tenant != nil {
+		delete(tenant.instances, instanceID)
+	}
 	ds.tenantsLock.Unlock()
 
 	// we may not have received any node stats for this instance
@@ -1073,6 +1080,49 @@ func (ds *Datastore) DeleteInstance(instanceID string) error {
 	msg := fmt.Sprintf("Deleted Instance %s", instanceID)
 	ds.db.logEvent(tenantID, string(userInfo), msg)
 
+	return nil
+}
+
+func (ds *Datastore) stopInstance(instanceID string) error {
+	ds.instancesLock.Lock()
+	i := ds.instances[instanceID]
+	oldNodeID := i.NodeID
+	i.NodeID = ""
+	i.State = payloads.Exited
+	ds.instancesLock.Unlock()
+
+	// we may not have received any node stats for this instance
+	if oldNodeID != "" {
+		ds.nodesLock.Lock()
+		delete(ds.nodes[oldNodeID].instances, instanceID)
+		ds.nodesLock.Unlock()
+	}
+
+	instanceStat := types.CiaoServerStats{
+		ID:        instanceID,
+		Timestamp: time.Now(),
+		Status:    payloads.Exited,
+	}
+	ds.instanceLastStatLock.Lock()
+	ds.instanceLastStat[instanceID] = instanceStat
+	ds.instanceLastStatLock.Unlock()
+
+	stats := []payloads.InstanceStat{
+		{
+			InstanceUUID: instanceID,
+			State:        payloads.Exited,
+		},
+	}
+
+	return errors.Wrapf(ds.db.addInstanceStats(stats, ""), "error adding instance stats to database")
+}
+
+// StopInstance removes the link between an instance and its node
+func (ds *Datastore) StopInstance(instanceID string) error {
+	err := ds.stopInstance(instanceID)
+	if err != nil {
+		return errors.Wrapf(err, "error stopping instance")
+	}
 	return nil
 }
